requirements.txt
flask
flask-socketio
flask-cors
pymongo
python-dotenv

